+++ src/components/maskedInput/MaskedInput.js
@@ -2,8 +2,8 @@ import _ from 'lodash';
 import PropTypes from 'prop-types';
 import React from 'react';
 import {StyleSheet, ViewPropTypes, Keyboard} from 'react-native';
-import BaseInput from './BaseInput';
-import TextField from './TextField';
+import BaseInput from '../baseInput/BaseInput';
+import TextField from '../textField/TextField';
 import View from '../view';
 import Text from '../text';
