+++ src/components/textField/TextField.js
@@ -10,9 +10,9 @@ import React from 'react';
 import {StyleSheet, Animated, TextInput as RNTextInput, Image as RNImage} from 'react-native';
 import {Constants} from '../../helpers';
 import {Colors, Typography, Spacings} from '../../style';
-import BaseInput from './BaseInput';
+import BaseInput from '../baseInput/BaseInput';
 import Modal from '../modal';
-import TextArea from './TextArea';
+import TextArea from '../textArea/TextArea';
 import View from '../view';
 import Image from '../image';
 import Text from '../text';
